Fix webrtc provider tests
