+++ homeassistant/components/go2rtc/__init__.py
@@ -128,6 +128,7 @@ class WebRTCProvider(CameraWebRTCProvider):
         self._rest_client = Go2RtcRestClient(self._session, url)
         self._sessions: dict[str, Go2RtcWsClient] = {}
 
+    @callback
     def async_is_supported(self, stream_source: str) -> bool:
         """Return if this provider is supports the Camera as source."""
         return stream_source.partition(":")[0] in _SUPPORTED_STREAMS
