+++ tests/components/camera/common.py
@@ -6,13 +6,6 @@ components. Instead call the service directly.
 
 from unittest.mock import Mock
 
-from homeassistant.components.camera import Camera
-from homeassistant.components.camera.webrtc import (
-    CameraWebRTCProvider,
-    async_register_webrtc_provider,
-)
-from homeassistant.core import HomeAssistant
-
 EMPTY_8_6_JPEG = b"empty_8_6"
 WEBRTC_ANSWER = "a=sendonly"
 STREAM_SOURCE = "rtsp://127.0.0.1/stream"
@@ -30,25 +23,3 @@ def mock_turbo_jpeg(
     mocked_turbo_jpeg.scale_with_quality.return_value = EMPTY_8_6_JPEG
     mocked_turbo_jpeg.encode.return_value = EMPTY_8_6_JPEG
     return mocked_turbo_jpeg
-
-
-async def add_webrtc_provider(hass: HomeAssistant) -> CameraWebRTCProvider:
-    """Add test WebRTC provider."""
-
-    class SomeTestProvider(CameraWebRTCProvider):
-        """Test provider."""
-
-        async def async_is_supported(self, stream_source: str) -> bool:
-            """Determine if the provider supports the stream source."""
-            return True
-
-        async def async_handle_web_rtc_offer(
-            self, camera: Camera, offer_sdp: str
-        ) -> str | None:
-            """Handle the WebRTC offer and return an answer."""
-            return "answer"
-
-    provider = SomeTestProvider()
-    async_register_webrtc_provider(hass, provider)
-    await hass.async_block_till_done()
-    return provider
